semiautonomous-agents/a2a-protocol-dojo/backend/pyproject.toml
[project]
name = "a2a-dojo-backend"
version = "0.1.0"
description = "A2A Protocol Dojo - Backend Gateway"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
